# perf: avoid signing transactions