### What I did

This makes testing faster as it skips signing.
Requires core PR: https://github.com/ApeWorX/ape/pull/2483

Honestly, didn't see as much of a perf gain as I hoped, but you do see some.
Command:

x3
```sh
ape test tests/test_provider.py::test_send_transaction
```

before:

```
1 passed in 0.46s
1 passed in 0.43s
1 passed in 0.43s
```

after:

```
1 passed in 0.43s
1 passed in 0.41s
1 passed in 0.40s
```

Tests will fail on the PR and lint, until the linked core PR is merged and available.

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)